Massive migration to use unique_ptr and true visitor pattern
I had mistakenly made everything a shared_ptr, not understanding the
ownership and lifetime of the different AST nodes. This commit fixes all
of that and adds some while statement and for statement handling to
allow for testing of a fibonacci test program.

With the migration to unique_ptr, we also had to migrate to a true
visitor pattern, as smart pointer casting does not work with unique_ptr