Function return statements workings
This commit is larger than ususally because I made a bunch of additions
to the code that would enable diagnosing code-path problems.

The underlying problem the caused this large excursion was that the the
unrolling of the stack pointer while catching the `Return` exception
object left the `m_current_environment` pointer of the `Interpreter`
class dangingly (pointing at nothing). This is addressed by explicitly
retoring the pointer when catching the result. There is probably a
better way to do this (for example, not using exceptions)...